requirements.txt
flask==2.3.3
pandas==2.1.1
numpy==1.24.3
psycopg2-binary==2.9.7
openai==0.28.1
cryptography==41.0.4
python-dotenv==1.0.0
requests==2.31.0
openpyxl==3.1.2
PyPDF2==3.0.1
plotly==5.16.1
scikit-learn==1.3.0